requirements.txt
# Production dependencies (none for this simple app)
google-genai>=1.0.0

# Development/testing dependencies
pytest>=7.0.0
build>=1.0.0

